[add]review

+++ db/schema.rb
@@ -133,7 +133,7 @@ ActiveRecord::Schema.define(version: 2020_05_09_081115) do
     t.integer "shop_id", null: false
     t.string "title", null: false
     t.text "body", null: false
-    t.integer "star", null: false
+    t.float "star", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
